Refactoring and styling
