+++ webui/src/pages/JobDetail/JobInfo/JobInfo.tsx
@@ -1,10 +1,12 @@
 import React from 'react';
-import { Job } from '../../../helpers/jobInterfaces';
-import { getShortenedJobID, fromTimestamp, capitalizeFirstLetter } from "../../../helpers/helperFunctions";
-
+import { Execution, Job } from '../../../helpers/jobInterfaces';
+import { fromTimestamp, capitalizeFirstLetter } from "../../../helpers/helperFunctions";
+import styles from "./JobInfo.module.scss";
+import Table from '../../../components/Table/Table';
 
 interface JobInfoProps {
   job: Job;
+  executions: Execution[];
   section: 'overview' | 'executionRecord' | 'executionDetails';
 }
 
@@ -13,13 +15,13 @@ interface DataItem {
     value: string | undefined;
 }  
 
-const JobInfo: React.FC<JobInfoProps> = ({ job, section }) => {
+const JobInfo: React.FC<JobInfoProps> = ({ job, executions, section }) => {
     let dataToDisplay: DataItem[] = [];
 
     switch (section) {
         case 'overview':
             dataToDisplay = [
-                { label: 'Job ID', value: getShortenedJobID(job.ID) },
+                { label: 'Job ID', value: job.ID },
                 { label: 'Job Type', value: capitalizeFirstLetter(job.Type) },
                 { label: 'Created', value: fromTimestamp(job.CreateTime).toString() },
                 { label: 'Modified', value: fromTimestamp(job.ModifyTime).toString() },
@@ -29,8 +31,8 @@ const JobInfo: React.FC<JobInfoProps> = ({ job, section }) => {
                 { label: 'Image', value: job.Tasks[0].Engine.Params.Image },
                 { label: 'GPU Details', value: job?.Tasks[0]?.Resources?.GPU ? job?.Tasks[0]?.Resources?.GPU : "Not specified" },
                 { label: 'Timeout', value: job?.Tasks[0]?.Timeouts.ExecutionTimeout.toString() },
-                { label: 'Requestor Node', value: 'Some value' },
-                { label: 'Concurrency', value: 'Some value' }
+                { label: 'Requestor Node', value: job.Meta["bacalhau.org/requester.id"]},
+                { label: 'Concurrency', value: '' }
             ];
             break;
         case 'executionRecord':
@@ -49,13 +51,40 @@ const JobInfo: React.FC<JobInfoProps> = ({ job, section }) => {
             break;
         default:
             break;
-    }      
+    }
+
+    const manyExecutions = executions.length > 1;
+    const tableData = {
+        headers: ["ID", "Created", "Modified", "Node ID", "Status", "Action"],
+        rows: executions.map(item => ({
+            "ID": item.ID,
+            "Created": fromTimestamp(item.CreateTime).toString(),
+            "Modified": fromTimestamp(item.ModifyTime).toString(),
+            "Node ID": item.NodeID,
+            "Status": capitalizeFirstLetter(item.DesiredState.Message),
+            "Action": <button onClick={() => handleShowClick(item)}>Show</button>
+        }))
+    };
+
+    const handleShowClick = (item: any) => {
+        // TODO: Logic to show selected execution
+        console.log('Showing details for:', item.ID);
+    };
 
     return (
-        <div>
-        {dataToDisplay.map(item => (
-            <p key={item.label}>{item.label}: {item.value}</p>
-        ))}
+        <div className={styles.jobInfo}>
+            {dataToDisplay.map(item => (
+                <p key={item.label} className={styles.item}>
+                    <span className={styles.key}>{item.label}: </span>
+                    <span className={styles.value}>{item.value}</span> 
+                </p>
+            ))}
+            {manyExecutions && section=='overview' && (
+              <div>
+                <span className={styles.key}>Executions List:</span>
+                    <Table data={tableData} style={{ fontSize: '12px' }}></Table>
+              </div>
+            )}
         </div>
     );
 };
